feat: ensure contents of container image changes
I got confused when the release tagging was being applied
to the same release package (container image).  This was because
the container image was not changing.

With this change, we put the contents of the repo into the
container image, ensuring there is change and a new image and digest
is created.  This will also cause new attestation to be created.

- [x] remove .github folder from .dockerignore to ensure the folder is put into the container image
- [x] update README explaining what the image contents are and why

Signed-off-by: jmeridth <jmeridth@gmail.com>